+++ src/main/java/kitchenpos/application/MenuService.java
@@ -3,11 +3,11 @@ package kitchenpos.application;
 import java.util.List;
 import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
-import kitchenpos.dao.MenuDao;
-import kitchenpos.dao.MenuGroupDao;
-import kitchenpos.dao.MenuProductDao;
-import kitchenpos.dao.ProductDao;
-import kitchenpos.domain.menu.Menu;
+import kitchenpos.dao.menu.MenuDao;
+import kitchenpos.dao.menugroup.MenuGroupDao;
+import kitchenpos.dao.menuproduct.MenuProductDao;
+import kitchenpos.dao.product.ProductDao;
+import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuProduct;
 import kitchenpos.domain.Product;
 import org.springframework.stereotype.Service;
