+++ src/main/java/kitchenpos/application/TableService.java
@@ -1,8 +1,8 @@
 package kitchenpos.application;
 
 import java.util.List;
-import kitchenpos.dao.OrderTableDao;
-import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.dao.ordertable.OrderTableDao;
+import kitchenpos.domain.OrderTable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
