+++ src/main/java/kitchenpos/dao/menu/MenuDao.java
@@ -1,6 +1,6 @@
-package kitchenpos.dao;
+package kitchenpos.dao.menu;
 
-import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.Menu;
 
 import java.util.List;
 import java.util.Optional;
