+++ src/main/java/kitchenpos/dao/order/OrderDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.dao.order;
 
 import kitchenpos.domain.Order;
