+++ src/main/java/kitchenpos/dao/product/ProductDao.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.dao.product;
 
 import kitchenpos.domain.Product;
