+++ src/test/java/kitchenpos/dao/JdbcTemplateMenuGroupDaoTest.java
@@ -4,6 +4,8 @@ import static kitchenpos.fixture.MenuFixture.getMenuGroupRequest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import kitchenpos.dao.menugroup.JdbcTemplateMenuGroupDao;
+import kitchenpos.dao.menugroup.MenuGroupDao;
 import kitchenpos.domain.MenuGroup;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
