+++ src/test/java/kitchenpos/dao/JdbcTemplateOrderDaoTest.java
@@ -8,6 +8,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Arrays;
 import java.util.List;
+import kitchenpos.dao.order.JdbcTemplateOrderDao;
+import kitchenpos.dao.order.OrderDao;
 import kitchenpos.domain.Order;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
