+++ src/test/java/kitchenpos/dao/JdbcTemplateOrderLineItemDaoTest.java
@@ -4,6 +4,10 @@ import static kitchenpos.fixture.OrderFixture.getOrderLineItem;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
+import kitchenpos.dao.order.JdbcTemplateOrderDao;
+import kitchenpos.dao.order.OrderDao;
+import kitchenpos.dao.orderlineitem.JdbcTemplateOrderLineItemDao;
+import kitchenpos.dao.orderlineitem.OrderLineItemDao;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.fixture.OrderFixture;
