+++ src/test/java/kitchenpos/dao/JdbcTemplateOrderTableDaoTest.java
@@ -3,7 +3,9 @@ package kitchenpos.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.domain.ordertable.OrderTable;
+import kitchenpos.dao.ordertable.JdbcTemplateOrderTableDao;
+import kitchenpos.dao.tablegroup.JdbcTemplateTableGroupDao;
+import kitchenpos.domain.OrderTable;
 import kitchenpos.fixture.TableFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
