+++ src/test/java/kitchenpos/dao/JdbcTemplateTableGroupDaoTest.java
@@ -3,7 +3,8 @@ package kitchenpos.dao;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.domain.tablegroup.TableGroup;
+import kitchenpos.dao.tablegroup.JdbcTemplateTableGroupDao;
+import kitchenpos.domain.TableGroup;
 import kitchenpos.fixture.TableFixture;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
